spliting AllProducts
